Updated MIDI timestamp from u32 to u64

+++ wasm_backend/src/framework/midi/MidiEvent.zig
@@ -10,13 +10,13 @@ pub fn noteNumberToFrequency(note_number: u32, frequency_of_a: f32) f32 {
 status: u8,
 d1: u8,
 d2: u8,
-sample_position: u32,
+sample_position: u64,
 
 pub fn asByte(v: u32) u8 {
     return @intCast(v & 0xff);
 }
 
-pub fn initFromPacked(packed_event: u32, sample_position: u32) @This() {
+pub fn initFromPacked(packed_event: u32, sample_position: u64) @This() {
     return .{
         .status = asByte(packed_event),
         .d1 = asByte(packed_event >> 8),
@@ -52,7 +52,7 @@ pub fn getVelocityFloat(self: @This()) f32 {
     return velocity_float * (1.0 / 127.0);
 }
 
-// Generel functions
+// General functions
 
 pub fn getSamplePosition(self: @This()) u32 {
     return self.sample_position;
